Add hierarchical trend display to choropleth map.
Updated the map code to provide to match the hierarchical display of
data in the prototype. At the national level, counties are displayed
(and colorized) on the map but only the state borders are highlighted.
At this level, the map callouts only show state level trends.

Upon selecting a state, the county borders for that state are "activated"
and within the selected state, map callouts will show the county level
statistics. Outside of the selected state, we continue to show state
level trends.

As part of this change, I went through and reorganized the code into
conceptual blocks dealing with various functions: drawing, event handling,
callouts, external interface. I also simplified the initialization sequence into
a single createMap call.