Fix integration issues with choropleth map.
Addressed several integration issues with the choropleth map that
is likely the underlying cause of several issues (#1, #18, #19).

1. The choropleth map expects the map to be created and initialized
prior to the setSelected* methods being called. The map is currently
initialized as part of the onMount method but the onChangeHandler is
getting called both before and after onMount. Added a boolean guard
to prevent the setSelected* methods from being called prior to the
maps initialization.

2. Found an issue where the TrendsOverview Svelte component was calling
the setSelectedState method with a sub_region_1_code instead of a
FIPS code to make the selection. Changed the interface to expect a
sub_region code instead of a FIPS code to integrate better.
(See issue #28.)