Update Next Steps to latest designs.
Updated the Next Steps section of the page to reflect the three column
design without icons. Added the third "next step" and removed the icon
place holder svgs and styling (Issue #26).

Note: this PR keeps the typography consistent with the current page
typography. Updates to track the latest design improvements to the
typographical styling will be addressed as a separate PR for the entire
page.