Change the interaction model for the map.
1. Change the cursor types used by the map to use the grab cursor any
time the cursor is within the bounds of the map SVG but not over a
selectable path and to use the pointer cursor when over a selectable
path of the map.

2. Added new pre-selection behavior to the map. When a user hovers over
a selectable region, that region will highlight via a stroke-width
increase until the user leaves the region.

3. Changed the map callout card behavior to display the card once a
user moves into a region and stops mousing for 200ms or more. Once the
card is displayed, it is "sticky" and stays displayed until the user
mouses out of the region. This applies a both the state and county
levels.

4. Changed the construction of the map callout card to use a static
layout with new CSS rules and simply modify the SVG colors and the
text of the value area to reflect the map values instead of creating
an SVG for the entire map info text.