Miscellaneous styling fixes from Tuesday RC review.
1. Change styling on .mapTrendSelector and .selectedTrend to change the
cursor to a pointer style when hovering over a selectable button and to
turn off pointer events for a non-selectable button.

2. Refactor construction of the map legend to use mostly HTML, instead
of drawing the legend as a large SVG in Javascript. Addressed styling
issues with the info button to have the entire bounding box of the
button targetable, not just the icon.

3. Added a hidden anchor to the About this data section. Changed the
navigation on the "Learn more" items to jump to this anchor.

4. Change text from "Safety and side effects searches" to
"Safety and side effect searches" (no plural on effect).

5. Fix styling of download information popup to ensure that all text
stays within the box. Reduced height of the box slightly to get better
layout.

6. Moved display of the map legend info popup into the TrendsOverview
component and refactored it to handle multiple info buttons.

7. Added info buttons and popups on each of the three trend charts.