Add date scrubber component to map controls.
Added a date scrubber component to the map controls to allow a user to
move the data display on the map back and forwards in time.

1. Updated the map header creation to include the date scrubber controls.

2. Improved the overall styling to be more conformant with the design
specifications for the header.

3. Replaced the info button svg with Material Icons font code to simplify
implementation.

4. Added a special-case for Alaska state borders. Alaska borders get
lost with a "negative space" border. Added a slight grey (grey-200)
border around Alaska to preserve shape across time.