Fix rounding issue in map callout cards.
The routine to display the trend values for the map callout card was
testing for the existing of a value using the single |. When a value
was present this caused it to be converted to an integer and bitwise
or-ed with 0. Changed the test to use the double || - logical or
instead of bitwise or.